Module wide::arch

Architecture specific functionality.

THIS MODULE IS AN IMPLEMENTATION DETAIL OF THE CRATE.

The API here is not subject to the semver promise of the crate and if you decide to directly use anything in here I am free to break your code.

Modules

x86_64
Intrinsics for the x86_64 processor family.

Structs

Align2
Wrap the inner value to a minimum alignment of 2.

Align4
Wrap the inner value to a minimum alignment of 4.

Align8
Wrap the inner value to a minimum alignment of 8.

Align16
Wrap the inner value to a minimum alignment of 16.

Align32
Wrap the inner value to a minimum alignment of 32.