Function wide::arch::x86_64::prefetch1

pub fn prefetch1(ptr: *const impl Sized)

Prefetch the cache line into L2 and higher.

A prefetch is just a hint to the CPU and has no effect on the correctness (or not) of a program. In other words, you can prefetch literally any address and it's never UB. However, if you prefetch an invalid address the CPU can actually slow down for a moment as it figures out that your address isn't valid. So, don't go silly with this.

See Also: _mm_prefetch