Struct wide::arch::x86_64::m128i

#[repr(transparent)]pub struct m128i(pub __m128i);

A 128-bit SIMD value. Integral data, lanes determined by each op.

This documentation numbers the lanes based on the index you'd need to use to access that lane if the value were cast to an array.

This is also the way that the type is printed out using Debug, Display, LowerExp, and UpperExp.

This is not necessarily the ordering you'll see if you look an xmm register in a debugger! Basically because of how little-endian works.

Most operations work per-lane, "lanewise".

Some operations work using lane 0 only. When appropriate, these have the same name as the lanewise version but with a 0 on the end. Eg: cmp_eq and cmp_eq0. The other lanes are simply copied forward from self.

Comparisons give "bool-ish" output, where all bits 1 in a lane is true, and all bits 0 in a lane is false. Unfortunately, all bits 1 with an f32 is one of the NaN values, and NaN != NaN, so it can be a little tricky to work with until you're used to it.

Implementations

impl m128i

#[must_use]pub fn add_i8(self, rhs: Self) -> Self

Lanewise i8 wrapping addition

#[must_use]pub fn add_i16(self, rhs: Self) -> Self

Lanewise i16 wrapping addition

#[must_use]pub fn add_i32(self, rhs: Self) -> Self

Lanewise i32 wrapping addition

#[must_use]pub fn add_i64(self, rhs: Self) -> Self

Lanewise i64 wrapping addition

#[must_use]pub fn saturating_add_i8(self, rhs: Self) -> Self

Lanewise i8 saturating addition

#[must_use]pub fn saturating_add_i16(self, rhs: Self) -> Self

Lanewise i16 saturating addition

#[must_use]pub fn saturating_add_u8(self, rhs: Self) -> Self

Lanewise u8 saturating addition

#[must_use]pub fn saturating_add_u16(self, rhs: Self) -> Self

Lanewise u16 saturating addition

#[must_use]pub fn andnot(self, rhs: Self) -> Self

Bitwise (!self) & rhs

#[must_use]pub fn average_u8(self, rhs: Self) -> Self

Lanewise u8 average: (a + b + 1) >> 1

#[must_use]pub fn average_u16(self, rhs: Self) -> Self

Lanewise u16 average: (a + b + 1) >> 1

#[must_use]pub fn cast_m128(self) -> m128

Cast the bits of this m128i directly to m128 without modification.

#[must_use]pub fn cmp_eq_i8(self, rhs: Self) -> Self

Lanewise i8 equality: bool-ish output

#[must_use]pub fn cmp_eq_i16(self, rhs: Self) -> Self

Lanewise i16 equality: bool-ish output

#[must_use]pub fn cmp_eq_i32(self, rhs: Self) -> Self

Lanewise i32 equality: bool-ish output

#[must_use]pub fn cmp_gt_i8(self, rhs: Self) -> Self

Lanewise i8 greater than: bool-ish output

#[must_use]pub fn cmp_gt_i16(self, rhs: Self) -> Self

Lanewise i16 greater than: bool-ish output

#[must_use]pub fn cmp_gt_i32(self, rhs: Self) -> Self

Lanewise i32 greater than: bool-ish output

#[must_use]pub fn cmp_lt_i8(self, rhs: Self) -> Self

Lanewise i8 greater than: bool-ish output

#[must_use]pub fn cmp_lt_i16(self, rhs: Self) -> Self

Lanewise i16 greater than: bool-ish output

#[must_use]pub fn cmp_lt_i32(self, rhs: Self) -> Self

Lanewise i32 greater than: bool-ish output

#[must_use]pub fn round_low_f64(self) -> m128d

Rounds the lower two i32 lanes to f64 lanes.

#[must_use]pub fn round_f32(self) -> m128

Rounds the i32 lanes to f32 lanes.

#[must_use]pub fn extract0_i32(self) -> i32

Gets out the lowest i32 lane.

#[must_use]pub fn extract0_i64(self) -> i64

Gets out the lowest i64 lane.

#[must_use]pub fn set0_i32(val: i32) -> Self

Places the i32 in the low lane and zeroes other lanes.

#[must_use]pub fn set0_i64(val: i64) -> Self

Places the i64 in the low lane and zeroes the other lane.

#[must_use]pub fn load(addr: &Align16<i128>) -> Self

Loads the aligned i128 address specified.

#[must_use]pub fn load0_i64(addr: &Align16<i64>) -> Self

Loads the aligned i64 address specified into the low lane.

#[must_use]pub fn load_unaligned(addr: &[u8; 16]) -> Self

Loads the i128 address specified, no alignment requirements.

#[must_use]pub fn mul_i16_hadd(self, rhs: Self) -> Self

Lanewise i16 multiply then horizontal add into four lanes.

The eight i16 multiplies produce eight intermediate i32 values, which then get horizontal added into four i32 values.

#[must_use]pub fn max_u8(self, rhs: Self) -> Self

Lanewise u8 maximum

#[must_use]pub fn min_u8(self, rhs: Self) -> Self

Lanewise u8 minimum

#[must_use]pub fn max_i16(self, rhs: Self) -> Self

Lanewise i16 maximum

#[must_use]pub fn min_i16(self, rhs: Self) -> Self

Lanewise i16 minimum

#[must_use]pub fn copy0_i64(self) -> Self

Copies the low lane i64 into a new value, upper lane is 0.

#[must_use]pub fn move_mask_i8(self) -> i32

Crates a move mask from the i8 lanes.

#[must_use]pub fn mul_high_i16(self, rhs: Self) -> Self

Lanewise i16 multiplication, keep high bits.

#[must_use]pub fn mul_low_i16(self, rhs: Self) -> Self

Lanewise i16 multiplication, keep low bits.

#[must_use]pub fn mul_high_u16(self, rhs: Self) -> Self

Lanewise u16 multiplication, keep high bits.

#[must_use]pub fn half_mul_u32(self, rhs: Self) -> Self

Lower half of each i64 lane is u32 multiplied into u64 lanes.

#[must_use]pub fn pack_i16_saturating_i8(self, rhs: Self) -> Self

Pack self then rhs i16 values into saturated i8s

#[must_use]pub fn pack_i32_saturating_i16(self, rhs: Self) -> Self

Pack self then rhs i32 values into saturated i16s

#[must_use]pub fn pack_i16_saturating_u8(self, rhs: Self) -> Self

Pack self then rhs i16 values into saturated u8s

#[must_use]pub fn signed_abs_diff_i8(self, rhs: Self) -> Self

Sum of absolute i8 differences, eight at a time into two i64 lanes.

#[must_use]pub fn set_i8(
a: i8,
b: i8,
c: i8,
d: i8,
e: i8,
f: i8,
g: i8,
h: i8,
i: i8,
j: i8,
k: i8,
l: i8,
m: i8,
n: i8,
o: i8,
p: i8
) -> Self

Sets the i8 values together in standard order.

#[must_use]pub fn set_reverse_i8(
a: i8,
b: i8,
c: i8,
d: i8,
e: i8,
f: i8,
g: i8,
h: i8,
i: i8,
j: i8,
k: i8,
l: i8,
m: i8,
n: i8,
o: i8,
p: i8
) -> Self

Sets the i8 values together in reverse order.

#[must_use]pub fn set_i16(
a: i16,
b: i16,
c: i16,
d: i16,
e: i16,
f: i16,
g: i16,
h: i16
) -> Self

Sets the i16 values together in standard order.

#[must_use]pub fn set_reverse_i16(
a: i16,
b: i16,
c: i16,
d: i16,
e: i16,
f: i16,
g: i16,
h: i16
) -> Self

Sets the i16 values together in reverse order.

#[must_use]pub fn set_i32(a: i32, b: i32, c: i32, d: i32) -> Self

Sets the i32 values together in standard order.

#[must_use]pub fn set_reverse_i32(a: i32, b: i32, c: i32, d: i32) -> Self

Sets the i32 values together in reverse order.

#[must_use]pub fn set_i64(a: i64, b: i64) -> Self

Sets the i64 values together in standard order.

#[must_use]pub fn splat_i8(a: i8) -> Self

Splats the i8 value across all lanes.

#[must_use]pub fn splat_i16(a: i16) -> Self

Splats the i16 value across all lanes.

#[must_use]pub fn splat_i32(a: i32) -> Self

Splats the i32 value across all lanes.

#[must_use]pub fn splat_i64(a: i64) -> Self

Splats the i64 value across all lanes.

#[must_use]pub fn shift_left_i16(self, rhs: Self) -> Self

Lanewise i16 left shift using rhs as a i128: self << rhs

#[must_use]pub fn shift_left_i32(self, rhs: Self) -> Self

Lanewise i32 left shift using rhs as a i128: self << rhs

#[must_use]pub fn shift_left_i64(self, rhs: Self) -> Self

Lanewise i64 left shift using rhs as a i128: self << rhs

#[must_use]pub fn shift_right_sign_i16(self, rhs: Self) -> Self

Lanewise i16 right shift using rhs as a i128: self >> rhs

Sign bit is preserved when shifting.

#[must_use]pub fn shift_right_sign_i32(self, rhs: Self) -> Self

Lanewise i32 right shift using rhs as a i128: self >> rhs

Sign bit is preserved when shifting.

#[must_use]pub fn shift_right_zero_i16(self, rhs: Self) -> Self

Lanewise i16 right shift using rhs as a i128: self >> rhs

Zeroes are shifted in regardless of the sign bit.

#[must_use]pub fn shift_right_zero_i32(self, rhs: Self) -> Self

Lanewise i32 right shift using rhs as a i128: self >> rhs

Zeroes are shifted in regardless of the sign bit.

#[must_use]pub fn shift_right_zero_i64(self, rhs: Self) -> Self

Lanewise i64 right shift using rhs as a i128: self >> rhs

Zeroes are shifted in regardless of the sign bit.

pub fn store(self, addr: &mut Align16<i128>)

Stores the data to the aligned address given.

pub fn store0_i64(self, addr: &mut Align16<i64>)

Stores the lower i64 lane to the aligned address given.

pub fn store_unaligned(self, addr: &mut [u8; 16])

Stores the data to the address given, no alignment requirements.

#[must_use]pub fn sub_i8(self, rhs: Self) -> Self

Lanewise i8 wrapping subtraction

#[must_use]pub fn sub_i16(self, rhs: Self) -> Self

Lanewise i16 wrapping subtraction

#[must_use]pub fn sub_i32(self, rhs: Self) -> Self

Lanewise i32 wrapping subtraction

#[must_use]pub fn sub_i64(self, rhs: Self) -> Self

Lanewise i64 wrapping subtraction

#[must_use]pub fn saturating_sub_i8(self, rhs: Self) -> Self

Lanewise i8 saturating subtraction

#[must_use]pub fn saturating_sub_i16(self, rhs: Self) -> Self

Lanewise i16 saturating subtraction

#[must_use]pub fn saturating_sub_u8(self, rhs: Self) -> Self

Lanewise u8 saturating subtraction

#[must_use]pub fn saturating_sub_u16(self, rhs: Self) -> Self

Lanewise u16 saturating subtraction

#[must_use]pub fn unpack_high_i8(self, rhs: Self) -> Self

Unpack i8 values from the high half of self and rhs

#[must_use]pub fn unpack_high_i16(self, rhs: Self) -> Self

Unpack i16 values from the high half of self and rhs

#[must_use]pub fn unpack_high_i32(self, rhs: Self) -> Self

Unpack i32 values from the high half of self and rhs

#[must_use]pub fn unpack_high_i64(self, rhs: Self) -> Self

Unpack i64 values from the high half of self and rhs

#[must_use]pub fn unpack_low_i8(self, rhs: Self) -> Self

Unpack i8 values from the low half of self and rhs

#[must_use]pub fn unpack_low_i16(self, rhs: Self) -> Self

Unpack i16 values from the low half of self and rhs

#[must_use]pub fn unpack_low_i32(self, rhs: Self) -> Self

Unpack i32 values from the low half of self and rhs

#[must_use]pub fn unpack_low_i64(self, rhs: Self) -> Self

Unpack i64 values from the low half of self and rhs

impl m128i

SSE3 Operations

#[must_use]pub fn load_quick_unaligned(addr: *const i128) -> Self

Loads 128-bits of integer data without alignment requirements.

This can perform faster than [m128i::load_unaligned] if the data would cross a cache line boundary.

impl m128i

SSSE3 Operations

#[must_use]pub fn abs_i8(self) -> Self

Lanewise i8 wrapping absolute value.

#[must_use]pub fn abs_i16(self) -> Self

Lanewise i16 wrapping absolute value.

#[must_use]pub fn abs_i32(self) -> Self

Lanewise i32 wrapping absolute value.

#[must_use]pub fn horizontal_add_i16(self, rhs: Self) -> Self

Horizontal add i16 pairs in self and rhs.

out[0]= self[1] + self[0]
out[1]= self[3] + self[2]
out[2]= self[5] + self[4]
out[3]= self[7] + self[6]
out[4]= rhs[1] + rhs[0]
out[5]= rhs[3] + rhs[2]
out[6]= rhs[5] + rhs[4]
out[7]= rhs[7] + rhs[6]

#[must_use]pub fn horizontal_saturating_add_i16(self, rhs: Self) -> Self

Horizontal saturating add i16 pairs in self and rhs.

out[0]= self[1].saturating_add(self[0])
out[1]= self[3].saturating_add(self[2])
out[2]= self[5].saturating_add(self[4])
out[3]= self[7].saturating_add(self[6])
out[4]= rhs[1].saturating_add(rhs[0])
out[5]= rhs[3].saturating_add(rhs[2])
out[6]= rhs[5].saturating_add(rhs[4])
out[7]= rhs[7].saturating_add(rhs[6])

#[must_use]pub fn horizontal_add_i32(self, rhs: Self) -> Self

Horizontal add i32 pairs in self and rhs.

out[0]= self[1] + self[0]
out[1]= self[3] + self[2]
out[2]= rhs[5] + rhs[4]
out[3]= rhs[7] + rhs[6]

#[must_use]pub fn horizontal_sub_i16(self, rhs: Self) -> Self

Horizontal subtract i16 pairs in self and rhs.

out[0]= self[0] - self[1]
out[1]= self[2] - self[3]
out[2]= self[4] - self[5]
out[3]= self[6] - self[7]
out[4]= rhs[0] - rhs[1]
out[5]= rhs[2] - rhs[3]
out[6]= rhs[4] - rhs[5]
out[7]= rhs[6] - rhs[7]

#[must_use]pub fn horizontal_saturating_sub_i16(self, rhs: Self) -> Self

Horizontal saturating subtract i16 pairs in self and rhs.

out[0]= self[1].saturating_sub(self[0])
out[1]= self[3].saturating_sub(self[2])
out[2]= self[5].saturating_sub(self[4])
out[3]= self[7].saturating_sub(self[6])
out[4]= rhs[1].saturating_sub(rhs[0])
out[5]= rhs[3].saturating_sub(rhs[2])
out[6]= rhs[5].saturating_sub(rhs[4])
out[7]= rhs[7].saturating_sub(rhs[6])

#[must_use]pub fn horizontal_sub_i32(self, rhs: Self) -> Self

Horizontal sub i32 pairs in self and rhs.

out[0]= self[1] - self[0]
out[1]= self[3] - self[2]
out[2]= rhs[5] - rhs[4]
out[3]= rhs[7] - rhs[6]

#[must_use]pub fn mul_hadd_u8_to_i16(self, rhs: Self) -> Self

Multiply u8 values in self and rhs into i16 intermediates, then horizontal add pack into the output.

out_i16[0]= self_u8[0]*rhs_u8[0] + self_u8[1]*rhs_u8[1]
out_i16[1]= self_u8[2]*rhs_u8[2] + self_u8[3]*rhs_u8[3]
out_i16[2]= self_u8[4]*rhs_u8[4] + self_u8[5]*rhs_u8[5]
out_i16[3]= self_u8[6]*rhs_u8[6] + self_u8[7]*rhs_u8[7]
out_i16[4]= self_u8[8]*rhs_u8[8] + self_u8[9]*rhs_u8[9]
out_i16[5]= self_u8[10]*rhs_u8[10] + self_u8[11]*rhs_u8[11]
out_i16[6]= self_u8[12]*rhs_u8[12] + self_u8[13]*rhs_u8[13]
out_i16[7]= self_u8[14]*rhs_u8[14] + self_u8[15]*rhs_u8[15]

#[must_use]pub fn mul_higher_ish_i16(self, rhs: Self) -> Self

Multiply i16 values in self and rhs into i32, keep high 18 bits, add 1, and then keep the middle 16 bits.

for i in 0..16 {
  out[i] = (((self[i] * rhs[i]) >> 14) + 1) >> 1
}

#[must_use]pub fn shuffle_i8(self, rhs: Self) -> Self

Shuffle i8 values in self according to control mask in rhs.

for i in 0..16 {
  out[i] = if rhs[i] < 0 {
    0
  } else {
    self[rhs[i] & 0b1111]
  };
}

#[must_use]pub fn sign_i8(self, rhs: Self) -> Self

i8: negate, zero, or no-change each lane of self based on rhs.

for i in 0..16 {
  out[i] = match sign(rhs[i]) {
    Positive => self[i],
    Zero => 0,
    Negative => -self[i],
  };
}

#[must_use]pub fn sign_i16(self, rhs: Self) -> Self

i16: negate, zero, or no-change each lane of self based on rhs.

for i in 0..8 {
  out[i] = match sign(rhs[i]) {
    Positive => self[i],
    Zero => 0,
    Negative => -self[i],
  };
}

#[must_use]pub fn sign_i32(self, rhs: Self) -> Self

i32: negate, zero, or no-change each lane of self based on rhs.

for i in 0..4 {
  out[i] = match sign(rhs[i]) {
    Positive => self[i],
    Zero => 0,
    Negative => -self[i],
  };
}