Struct wide::f32x4

#[repr(C, align(16))]pub struct f32x4 { /* fields omitted */ }

Four f32 values packed together.

Implementations

impl f32x4

Wide Methods

#[must_use]pub fn andnot(self, rhs: Self) -> Self

Bitwise (!self) & rhs

#[must_use]pub fn is_nan(self) -> Self

Lanewise "is the value a NaN value?"

#[must_use]pub fn is_ordinary(self) -> Self

Lanewise "is the value a non-NaN value?"

#[must_use]pub fn merge(self, a: Self, b: Self) -> Self

Use self (a boolish value) to merge a and b.

For each lane index, if the self lane is "true" then the a value will be used in the output, otherwise the b value will be used in the output.

If sse4.1 is enabled, then "true" only checks the sign bit. With less features enabled the entire bit pattern of the lane will matter. This is not normally a problem, because the comparison methods naturally return all 1s or all 0s anyway.

#[must_use]pub fn move_mask(self) -> i32

use wide::f32x4;
let a = f32x4::new(1.0, 2.0, 3.0, 4.0);
let b = f32x4::from(2.5);
assert_eq!(a.cmp_lt(b).move_mask(), 0b0011);

#[must_use]pub fn cmp_eq(self, rhs: Self) -> Self

Lanewise ==

#[must_use]pub fn cmp_ge(self, rhs: Self) -> Self

Lanewise >=

#[must_use]pub fn cmp_gt(self, rhs: Self) -> Self

Lanewise >

#[must_use]pub fn cmp_le(self, rhs: Self) -> Self

Lanewise <=

#[must_use]pub fn cmp_lt(self, rhs: Self) -> Self

Lanewise <

#[must_use]pub fn cmp_nan(self, rhs: Self) -> Self

Lanewise "self.is_nan() | other.is_nan()"

#[must_use]pub fn cmp_ne(self, rhs: Self) -> Self

Lanewise !=

#[must_use]pub fn cmp_nge(self, rhs: Self) -> Self

Lanewise !(a >= b)

If you call this method it triggers Third Impact

#[must_use]pub fn cmp_ngt(self, rhs: Self) -> Self

Lanewise !(a > b)

#[must_use]pub fn cmp_nle(self, rhs: Self) -> Self

Lanewise !(a <= b)

#[must_use]pub fn cmp_nlt(self, rhs: Self) -> Self

Lanewise !(a < b)

#[must_use]pub fn ceil(self) -> Self

Lanewise ceiling (next larger whole number)

#[must_use]pub fn floor(self) -> Self

Lanewise floor (next smaller whole number)

#[must_use]pub fn abs(self) -> Self

Lanewise absolute value

#[must_use]pub fn cos(self) -> Self

Lanewise cosine

#[must_use]pub fn round(self) -> Self

Lanewise round to nearest whole number.

#[must_use]pub fn sin(self) -> Self

Lanewise sine.

"We called it 'Sin'."

#[must_use]pub fn mul_add(self, b: Self, c: Self) -> Self

Performs a "fused multiply-add", (self * b) + c

This is only different from a normal mul and then add if the fma target_feature is enabled during the build. This is not on by default, you must enable it yourself. To be clear, this is not a cargo feature, this is a CPU feature. Read the guide for info on how to enable CPU features if you haven't done that before.

#[must_use]pub fn negated_mul_add(self, b: Self, c: Self) -> Self

Negated mul_add, c - (self * b).

Fused if fma feature is enabled, see mul_add.

#[must_use]pub fn recip(self) -> Self

Lanewise reciprocal, 1.0/lane

#[must_use]pub fn max(self, b: Self) -> Self

Lanewise maximum

#[must_use]pub fn min(self, b: Self) -> Self

Lanewise minimum

#[must_use]pub fn round_i32(self) -> i32x4

Lanewise round and get i32x4 output

#[must_use]pub fn is_finite(self) -> f32x4

If it's some finite value.

True for normal, denormal, and zero.

False for +/- INF, NaN

boolish

#[must_use]pub fn cast_i32x4(self) -> i32x4

Cast directly to i32x4 (no change in bit pattern).

#[must_use]pub fn copysign(self, b: Self) -> Self

Copy the sign of b onto the value of self.

#[must_use]pub fn clamp(self, min: Self, max: Self) -> Self

Lanewise clamp to a per-lane min and max.

#[must_use]pub fn signum(self) -> Self

Lanewise sign-number (-1.0 if < 0.0, otherwise 1.0)

#[must_use]pub fn tan(self) -> Self

Lanewise tangent

#[must_use]pub fn sin_cos(self) -> (Self, Self)

Sine and Cosine as a single operation.

#[must_use]pub fn to_degrees(self) -> Self

Lanewise radians -> degrees.

#[must_use]pub fn to_radians(self) -> Self

Lanewise degrees -> radians.

#[must_use]pub fn fract(self) -> Self

Lanewise fractional part.

#[must_use]pub fn sqrt(self) -> Self

Lanewise sqrt

impl f32x4

pub const EPSILON: f32x4

Machine epsilon value for f32.

pub const INFINITY: f32x4

Positive Infinity (∞).

pub const MAX: f32x4

Largest finite f32 value.

pub const MIN: f32x4

Smallest finite f32 value.

pub const MIN_POSITIVE: f32x4

Smallest positive normal f32 value.

pub const NAN: f32x4

Not a Number (NaN).

Reminder: This is one possible NaN value, but there are many NaN bit patterns within the f32 space.

pub const NEG_INFINITY: f32x4

Negative infinity (-∞).

pub const E: f32x4

Euler's number (e)

pub const FRAC_1_PI: f32x4

1/π

pub const FRAC_2_PI: f32x4

2/π

pub const FRAC_2_SQRT_PI: f32x4

2/sqrt(π)

pub const FRAC_1_SQRT_2: f32x4

1/sqrt(2)

pub const FRAC_PI_2: f32x4

π/2

pub const FRAC_PI_3: f32x4

π/3

pub const FRAC_PI_4: f32x4

π/4

pub const FRAC_PI_6: f32x4

π/6

pub const FRAC_PI_8: f32x4

π/8

pub const LN_2: f32x4

ln(2)

pub const LN_10: f32x4

ln(10)

pub const LOG2_E: f32x4

log2(e)

pub const LOG10_E: f32x4

log10(e)

pub const PI: f32x4

Archimedes' constant (π)

pub const SQRT_2: f32x4

sqrt(2)

pub const ALL_BITS_ACTIVE: f32x4

All bits active.

pub const ALL_EXCEPT_SIGN: f32x4

All bits active.

pub const ZERO: f32x4

0.0

pub const NEGATIVE_ZERO: f32x4

-0.0

pub const HALF: f32x4

0.5

pub const ONE: f32x4

1.0

pub const NEGATIVE_ONE: f32x4

1.0

pub const TWO_PI: f32x4

2.0 * π, the number of radians in a circle.

impl f32x4

#[must_use]pub fn new(a: f32, b: f32, c: f32, d: f32) -> Self

Makes a new f32x4.

impl f32x4

pub fn exp2(self) -> Self

pub fn exp(self) -> Self

pub fn log10(self) -> Self

pub fn log2(self) -> Self

#[must_use]pub fn trunc(self) -> Self

Truncate the fractional part.

#[must_use]pub fn ln(self) -> Self

pub fn powf(self, b: Self) -> Self

pub fn powi(self, b: [i32; 4]) -> Self

#[must_use]pub fn classify(self) -> [FpCategory; 4]

As [f32::classify]

#[must_use]pub fn acos(self) -> Self

#[must_use]pub fn acosh(self) -> Self

pub fn asin(self) -> Self

pub fn asinh(self) -> Self

pub fn atan(self) -> Self

pub fn atanh(self) -> Self

pub fn cbrt(self) -> Self

pub fn cosh(self) -> Self

pub fn exp_m1(self) -> Self

pub fn ln_1p(self) -> Self

pub fn log(self, b: Self) -> Self

pub fn sinh(self) -> Self

pub fn tanh(self) -> Self

pub fn atan2(self, b: Self) -> Self

pub fn hypot(self, b: Self) -> Self