Struct wide::f64x2

#[repr(C, align(16))]
pub struct f64x2 { /* private fields */ }

Implementations

impl f64x2

pub const ONE: f64x2 = unsafe { ConstUnionHack128bit{f64a2: [1.0; 2],}.f64x2 }

pub const ZERO: f64x2 = unsafe { ConstUnionHack128bit{f64a2: [0.0; 2],}.f64x2 }

pub const HALF: f64x2 = unsafe { ConstUnionHack128bit{f64a2: [0.5; 2],}.f64x2 }

pub const E: f64x2 = unsafe { ConstUnionHack128bit{f64a2: [core::f64::consts::E; 2],}.f64x2 }

pub const FRAC_1_PI: f64x2 = unsafe { ConstUnionHack128bit{f64a2: [core::f64::consts::FRAC_1_PI; 2],}.f64x2 }

pub const FRAC_2_PI: f64x2 = unsafe { ConstUnionHack128bit{f64a2: [core::f64::consts::FRAC_2_PI; 2],}.f64x2 }

pub const FRAC_2_SQRT_PI: f64x2 = unsafe { ConstUnionHack128bit{f64a2: [core::f64::consts::FRAC_2_SQRT_PI; 2],}.f64x2 }

pub const FRAC_1_SQRT_2: f64x2 = unsafe { ConstUnionHack128bit{f64a2: [core::f64::consts::FRAC_1_SQRT_2; 2],}.f64x2 }

pub const FRAC_PI_2: f64x2 = unsafe { ConstUnionHack128bit{f64a2: [core::f64::consts::FRAC_PI_2; 2],}.f64x2 }

pub const FRAC_PI_3: f64x2 = unsafe { ConstUnionHack128bit{f64a2: [core::f64::consts::FRAC_PI_3; 2],}.f64x2 }

pub const FRAC_PI_4: f64x2 = unsafe { ConstUnionHack128bit{f64a2: [core::f64::consts::FRAC_PI_4; 2],}.f64x2 }

pub const FRAC_PI_6: f64x2 = unsafe { ConstUnionHack128bit{f64a2: [core::f64::consts::FRAC_PI_6; 2],}.f64x2 }

pub const FRAC_PI_8: f64x2 = unsafe { ConstUnionHack128bit{f64a2: [core::f64::consts::FRAC_PI_8; 2],}.f64x2 }

pub const LN_2: f64x2 = unsafe { ConstUnionHack128bit{f64a2: [core::f64::consts::LN_2; 2],}.f64x2 }

pub const LN_10: f64x2 = unsafe { ConstUnionHack128bit{f64a2: [core::f64::consts::LN_10; 2],}.f64x2 }

pub const LOG2_E: f64x2 = unsafe { ConstUnionHack128bit{f64a2: [core::f64::consts::LOG2_E; 2],}.f64x2 }

pub const LOG10_E: f64x2 = unsafe { ConstUnionHack128bit{f64a2: [core::f64::consts::LOG10_E; 2],}.f64x2 }

pub const LOG10_2: f64x2 = unsafe { ConstUnionHack128bit{f64a2: [core::f64::consts::LOG10_2; 2],}.f64x2 }

pub const LOG2_10: f64x2 = unsafe { ConstUnionHack128bit{f64a2: [core::f64::consts::LOG2_10; 2],}.f64x2 }

pub const PI: f64x2 = unsafe { ConstUnionHack128bit{f64a2: [core::f64::consts::PI; 2],}.f64x2 }

pub const SQRT_2: f64x2 = unsafe { ConstUnionHack128bit{f64a2: [core::f64::consts::SQRT_2; 2],}.f64x2 }

pub const TAU: f64x2 = unsafe { ConstUnionHack128bit{f64a2: [core::f64::consts::TAU; 2],}.f64x2 }

impl f64x2

pub fn new(array: [f64; 2]) -> Self

pub fn blend(self, t: Self, f: Self) -> Self

pub fn abs(self) -> Self

pub fn fast_max(self, rhs: Self) -> Self

Calculates the lanewise maximum of both vectors. This is a faster implementation than max, but it doesn’t specify any behavior if NaNs are involved.

pub fn max(self, rhs: Self) -> Self

Calculates the lanewise maximum of both vectors. If either lane is NaN, the other lane gets chosen. Use fast_max for a faster implementation that doesn’t handle NaNs.

pub fn fast_min(self, rhs: Self) -> Self

Calculates the lanewise minimum of both vectors. This is a faster implementation than min, but it doesn’t specify any behavior if NaNs are involved.

pub fn min(self, rhs: Self) -> Self

Calculates the lanewise minimum of both vectors. If either lane is NaN, the other lane gets chosen. Use fast_min for a faster implementation that doesn’t handle NaNs.

pub fn is_nan(self) -> Self

pub fn is_finite(self) -> Self

pub fn is_inf(self) -> Self

pub fn round(self) -> Self

pub fn round_int(self) -> i64x2

pub fn mul_add(self, m: Self, a: Self) -> Self

pub fn mul_sub(self, m: Self, a: Self) -> Self

pub fn mul_neg_add(self, m: Self, a: Self) -> Self

pub fn mul_neg_sub(self, m: Self, a: Self) -> Self

pub fn flip_signs(self, signs: Self) -> Self

pub fn copysign(self, sign: Self) -> Self

pub fn asin_acos(self) -> (Self, Self)

pub fn acos(self) -> Self

pub fn asin(self) -> Self

pub fn atan(self) -> Self

pub fn atan2(self, x: Self) -> Self

pub fn sin_cos(self) -> (Self, Self)

pub fn sin(self) -> Self

pub fn cos(self) -> Self

pub fn tan(self) -> Self

pub fn to_degrees(self) -> Self

pub fn to_radians(self) -> Self

pub fn sqrt(self) -> Self

pub fn move_mask(self) -> i32

pub fn any(self) -> bool

pub fn all(self) -> bool

pub fn none(self) -> bool

pub fn exp(self) -> Self

Calculate the exponent of a packed f64x2

pub fn reduce_add(self) -> f64

pub fn ln(self) -> Self

pub fn log2(self) -> Self

pub fn log10(self) -> Self

pub fn pow_f64x2(self, y: Self) -> Self

pub fn powf(self, y: f64) -> Self

pub fn to_array(self) -> [f64; 2]

pub fn as_array_ref(&self) -> &[f64; 2]

impl f64x2

pub fn splat(elem: f64) -> f64x2