Struct wide::f32x8

#[repr(C, align(32))]
pub struct f32x8 { /* fields omitted */ }

Implementations

impl f32x8

pub const ONE: f32x8

pub const HALF: f32x8

pub const ZERO: f32x8

pub const E: f32x8

pub const FRAC_1_PI: f32x8

pub const FRAC_2_PI: f32x8

pub const FRAC_2_SQRT_PI: f32x8

pub const FRAC_1_SQRT_2: f32x8

pub const FRAC_PI_2: f32x8

pub const FRAC_PI_3: f32x8

pub const FRAC_PI_4: f32x8

pub const FRAC_PI_6: f32x8

pub const FRAC_PI_8: f32x8

pub const LN_2: f32x8

pub const LN_10: f32x8

pub const LOG2_E: f32x8

pub const LOG10_E: f32x8

pub const LOG10_2: f32x8

pub const LOG2_10: f32x8

pub const PI: f32x8

pub const SQRT_2: f32x8

pub const TAU: f32x8

impl f32x8

pub fn blend(self, t: Self, f: Self) -> Self

pub fn abs(self) -> Self

pub fn fast_max(self, rhs: Self) -> Self

Calculates the lanewise maximum of both vectors. This is a faster implementation than max, but it doesn’t specify any behavior if NaNs are involved.

pub fn max(self, rhs: Self) -> Self

Calculates the lanewise maximum of both vectors. This doesn’t match IEEE-754 and instead is defined as self < rhs ? rhs : self.

pub fn fast_min(self, rhs: Self) -> Self

Calculates the lanewise minimum of both vectors. This is a faster implementation than min, but it doesn’t specify any behavior if NaNs are involved.

pub fn min(self, rhs: Self) -> Self

Calculates the lanewise minimum of both vectors. If either lane is NaN, the other lane gets chosen. Use fast_min for a faster implementation that doesn’t handle NaNs.

pub fn is_nan(self) -> Self

pub fn is_finite(self) -> Self

pub fn is_inf(self) -> Self

pub fn round(self) -> Self

pub fn fast_round_int(self) -> i32x8

Rounds each lane into an integer. This is a faster implementation than round_int, but it doesn’t handle out of range values or NaNs. For those values you get implementation defined behavior.

pub fn round_int(self) -> i32x8

Rounds each lane into an integer. This saturates out of range values and turns NaNs into 0. Use fast_round_int for a faster implementation that doesn’t handle out of range values or NaNs.

pub fn fast_trunc_int(self) -> i32x8

Truncates each lane into an integer. This is a faster implementation than trunc_int, but it doesn’t handle out of range values or NaNs. For those values you get implementation defined behavior.

pub fn trunc_int(self) -> i32x8

Truncates each lane into an integer. This saturates out of range values and turns NaNs into 0. Use fast_trunc_int for a faster implementation that doesn’t handle out of range values or NaNs.

pub fn mul_add(self, m: Self, a: Self) -> Self

pub fn mul_sub(self, m: Self, a: Self) -> Self

pub fn mul_neg_add(self, m: Self, a: Self) -> Self

pub fn mul_neg_sub(self, m: Self, a: Self) -> Self

pub fn flip_signs(self, signs: Self) -> Self

pub fn copysign(self, sign: Self) -> Self

pub fn asin_acos(self) -> (Self, Self)

pub fn asin(self) -> Self

pub fn acos(self) -> Self

pub fn atan(self) -> Self

pub fn atan2(self, x: Self) -> Self

pub fn sin_cos(self) -> (Self, Self)

pub fn sin(self) -> Self

pub fn cos(self) -> Self

pub fn tan(self) -> Self

pub fn to_degrees(self) -> Self

pub fn to_radians(self) -> Self

pub fn recip(self) -> Self

pub fn recip_sqrt(self) -> Self

pub fn sqrt(self) -> Self

pub fn move_mask(self) -> i32

pub fn any(self) -> bool

pub fn all(self) -> bool

pub fn none(self) -> bool

pub fn exp(self) -> Self

Calculate the exponent of a packed f32x8

pub fn sign_bit(self) -> Self

pub fn reduce_add(self) -> f32

pub fn ln(self) -> Self

Natural log (ln(x))

pub fn log2(self) -> Self

pub fn log10(self) -> Self

pub fn pow_f32x8(self, y: Self) -> Self

pub fn powf(self, y: f32) -> Self

pub fn to_array(self) -> [f32; 8]

pub fn as_array_ref(&self) -> &[f32; 8]

impl f32x8

pub fn splat(elem: f32) -> f32x8