wide
Struct i16x8

#[repr(C, align(16))]
pub struct i16x8 { /* private fields */ }

Implementations

impl i16x8

pub const ONE: i16x8 = _

pub const ZERO: i16x8 = _

pub const MAX: i16x8 = _

pub const MIN: i16x8 = _

pub const LANES: u16 = 8u16

The number of lanes in this SIMD vector.

pub const BITS: u16 = 128u16

The size of this SIMD vector in bits.

impl i16x8

pub const fn new(array: [i16; 8]) -> Self

pub fn move_mask(self) -> i32

pub fn any(self) -> bool

pub fn all(self) -> bool

pub fn none(self) -> bool

pub fn from_u8x16_low(u: u8x16) -> Self

Unpack the lower half of the input and expand it to i16 values.

pub fn from_u8x16_high(u: u8x16) -> Self

Unpack the upper half of the input and expand it to i16 values.

pub fn from_i32x8_saturate(v: i32x8) -> Self

returns low i16 of i32, saturating values that are too large

pub fn from_i32x8_truncate(v: i32x8) -> Self

returns low i16 of i32, truncating the upper bits if they are set

pub fn from_slice_unaligned(input: &[i16]) -> Self

pub fn blend(self, t: Self, f: Self) -> Self

pub fn is_negative(self) -> Self

pub fn reduce_add(self) -> i16

horizontal add of all the elements of the vector

pub fn reduce_min(self) -> i16

horizontal min of all the elements of the vector

pub fn reduce_max(self) -> i16

horizontal max of all the elements of the vector

pub fn abs(self) -> Self

pub fn unsigned_abs(self) -> u16x8

pub fn max(self, rhs: Self) -> Self

pub fn min(self, rhs: Self) -> Self

pub fn saturating_add(self, rhs: Self) -> Self

pub fn saturating_sub(self, rhs: Self) -> Self

pub fn dot(self, rhs: Self) -> i32x4

Calculates partial dot product. Multiplies packed signed 16-bit integers, producing intermediate signed 32-bit integers. Horizontally add adjacent pairs of intermediate 32-bit integers.

pub fn mul_scale_round(self, rhs: Self) -> Self

Multiply and scale equivalent to ((self * rhs) + 0x4000) >> 15 on each lane, effectively multiplying by a 16 bit fixed point number between -1 and 1. This corresponds to the following instructions:

vqrdmulhq_s16 instruction on neon

i16x8_q15mulr_sat on simd128

_mm_mulhrs_epi16 on ssse3

emulated via mul_i16_* on sse2

pub fn mul_keep_high(lhs: Self, rhs: Self) -> Self

Multiples two i16x8 and return the high part of intermediate i32x8

pub fn mul_widen(self, rhs: Self) -> i32x8

multiplies two i16x8 and returns the result as a widened i32x8

pub fn transpose(data: [i16x8; 8]) -> [i16x8; 8]

transpose matrix of 8x8 i16 matrix

pub fn mul_scale_round_n(self, rhs: i16) -> Self

Multiply and scale, equivalent to ((self * rhs) + 0x4000) >> 15 on each lane, effectively multiplying by a 16 bit fixed point number between -1 and 1. This corresponds to the following instructions:

vqrdmulhq_n_s16 instruction on neon

i16x8_q15mulr_sat on simd128

_mm_mulhrs_epi16 on ssse3

emulated via mul_i16_* on sse2

pub fn to_array(self) -> [i16; 8]

pub fn as_array_ref(&self) -> &[i16; 8]

pub fn as_array_mut(&mut self) -> &mut [i16; 8]

impl i16x8

pub fn splat(elem: i16) -> i16x8